% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recpd_cor_calc.R
\name{recpd_cor_calc}
\alias{recpd_cor_calc}
\title{RecPD branch-length weighted jaccard similarity.}
\usage{
recpd_cor_calc(res, feats = NA)
}
\arguments{
\item{res}{A results data.frame generated by recpd_calc().}

\item{feats}{Optional: A user-provided array of at least length 2, containing named features, or indices,
corresponding to the feature column/rows of the recpd_calc() results data.frame.}
}
\value{
data.frame of feature x feature RecPD correlations.
}
\description{
RecPD branch-length weighted jaccard similarity.
}
\examples{

}
