% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recpd_calc.R
\name{recpd_calc}
\alias{recpd_calc}
\title{Calculate the RecPD of a feature and other associated measures.}
\usage{
recpd_calc(
  tree,
  tab,
  option = c("nn", "mpr", "ace"),
  params = list(nn_select = c("min", "median"), model_type = c("ER", "ARD"), kappa = 1,
    lik_cut = 0.99),
  calc = FALSE,
  prune = TRUE
)
}
\arguments{
\item{tree}{A phylogenetic tree, in phylo format.}

\item{tab}{A feature(s) presence/absence array, matrix, or data.frame:
columns should be named according to the tip labels of the tree.}

\item{option}{Ancestral state reconstruction method to apply, either 'nn',
'mpr', or 'ace'.}

\item{params}{Additional paramaters to pass to ancestral reconstruction
methods.}

\item{calc}{logical. If TRUE, calculate additional measures: span,
clustering, longevity, and lability.}

\item{prune}{logical. If TRUE and option == 'nn', will provide a conservative
reconstruction of feature lineages.}
}
\value{
A named list including metrics and ancestral state annotations for
tree nodes and branches for each feature in tab.
}
\description{
See \href{https://www.biorxiv.org/content/10.1101/2021.10.01.462747v1}{Bundalovic-Torma, Guttman; (2021).}
}
\examples{

}
\references{
Insert reference with Rdpack.
}
\author{
Cedoljub Bundalovic-Torma
}
